Add many new features, such as dmaext, individual file support, and cleanup

+++ src/file.h
@@ -5,10 +5,10 @@
 unsigned file_size(const char *fn);
 
 /* load a file into an existing buffer */
-void *file_load_into(const char *fn, unsigned *sz, void *dst);
+void *file_load_into(const char *fn, size_t *sz, void *dst);
 
 /* load a file */
-void *file_load(const char *fn, unsigned *sz);
+void *file_load(const char *fn, size_t *sz);
 
 /* write file */
 unsigned file_write(const char *fn, void *data, unsigned data_sz);
